webrtc/README: Update instructions and use localhost
Do not recommend changing internal Chrome flags. This is very unsecure
and could open the door to attacks. Instead, web browsers consider
"localhost" as a privileged exception to the Secure Context rule, which
means that it can be used without SSL to access the mic.

Also add instructions for running with HTTPS, in case someone wants to
test the demo in their LAN instead of localhost.