webrtc/server: Fix asyncio event loop reference
The `get_event_loop()` call was creating a top-level loop which is not
what is intended by the code in `__run_audio_xfer()`, which runs already
as a Task with its own event loop, so `get_running_loop()` should be
used instead, inside the Task. Also, `get_event_loop()` is deprecated.