webrtc: Several fixes for performance and user friendliness (#167)
* webrtc/README: Update instructions and use localhost

Do not recommend changing internal Chrome flags. This is very unsecure
and could open the door to attacks. Instead, web browsers consider
"localhost" as a privileged exception to the Secure Context rule, which
means that it can be used without SSL to access the mic.

Also add instructions for running with HTTPS, in case someone wants to
test the demo in their LAN instead of localhost.

* webrtc/server: Add VOSK_KEY_FILE to pass key.pem file

Allows using typical cert/key certificate files, such as those generated
by mkcert.

* webrtc/server: Fix asyncio event loop reference

The `get_event_loop()` call was creating a top-level loop which is not
what is intended by the code in `__run_audio_xfer()`, which runs already
as a Task with its own event loop, so `get_running_loop()` should be
used instead, inside the Task. Also, `get_event_loop()` is deprecated.

* webrtc/server: Improved handling of AcceptWaveform() results

Make `process_chunk()` return None if an error is raised by the Python
layer, or if an empty PartialResult is returned. This avoids lots of
redundant DataChannel messages with empty payload.

* webrtc/server: Minor optimizations

* `max_frames_len` never changes, so no need to update it on every loop
iteration.

* Send an empty object to the HTML UI, so the user knows when
DataChannel has been connected and the demo is ready.

* webrtc/static: Show a prompt (`> `) on the "partial" div

Improve visual feedback to the user, by showing a small prompt indicator
on the div where "partial" messagess will appear. This gives the feeling
that the partial line does indeed contain intermediate results.

* webrtc/static: Improve DataChannel handling code

Catch errors, rename Text result div to "text", and use more robut
checks on object members.

* webrtc/static: Disable favicon to avoid errors on page loading